activePar {maxLik}	R Documentation

free parameters under maximization

Description

Return a logical vector, indicating which parameters were free under maximization, as opposed to the fixed parameters that are treated as constants. See argument “fixed” for maxNR.

Usage

activePar(x, ...)
## Default S3 method:
activePar(x, ...)

Arguments

x
object, created by a maximization routine, such as maxNR or maxLik, or derived from a maximization object.

...
further arguments for methods

Details

Several optimization routines allow the user to fix some parameter values (or do it automatically in some cases). For gradient or Hessian based inference one has to know which parameters carry optimization-related information.

Value

A logical vector, indicating whether the parameters were free to change during optimization algorithm.

Author(s)

Ott Toomet

See Also

maxNR, nObs

Examples

## a two-dimensional exponential hat
f <- function(a) exp(-a[1]^2 - a[2]^2)

## maximize wrt. both parameters 
free <- maxNR(f, start=1:2) 
summary(free)  # results should be close to (0,0)
activePar(free)

## keep the first parameter constant
cons <- maxNR(f, start=1:2, fixed=c(TRUE,FALSE))
summary(cons) # result should be around (1,0)
activePar(cons)

[Package maxLik version 1.5-2.2 Index]