maxLik-internal {maxLik}	R Documentation

Internal maxLik Functions

Description

Internal maxLik Functions

Details

These are either various methods, or functions, not intended to be called directly by the user (or in some cases are just waiting for proper documentation to be written :).

[Package maxLik version 1.5-2.2 Index]