AIC.maxLik {maxLik}	R Documentation

Methods for the various standard functions

Description

These are methods for the maxLik related objects. See also the documentation for the corresponding generic functions

Usage

## S3 method for class 'maxLik'
AIC(object, ..., k=2)
## S3 method for class 'maxim'
coef(object, ...)
## S3 method for class 'maxLik'
coef(object, ...)
## S3 method for class 'maxLik'
stdEr(x, eigentol=1e-12, ...)

Arguments

object
a ‘maxLik’ object (coef can also handle ‘maxim’ objects)

k
numeric, the penalty per parameter to be used; the default ‘k = 2’ is the classical AIC.

x
a ‘maxLik’ object

eigentol
The standard errors are only calculated if the ratio of the smallest and largest eigenvalue of the Hessian matrix is less than “eigentol”. Otherwise the Hessian is treated as singular.

...
other arguments for methods

Details

AIC

calculates Akaike's Information Criterion (and other information criteria).

coef

extracts the estimated parameters (model's coefficients).

stdEr

extracts standard errors (using the Hessian matrix).

Examples

## estimate mean and variance of normal random vector
set.seed(123)
x <- rnorm(50, 1, 2)

## log likelihood function.
## Note: 'param' is a vector
llf <- function( param ) {
   mu <- param[ 1 ]
   sigma <- param[ 2 ]
   return(sum(dnorm(x, mean=mu, sd=sigma, log=TRUE)))
}

## Estimate it.  Take standard normal as start values
ml <- maxLik(llf, start = c(mu=0, sigma=1) )

coef(ml)
stdEr(ml)
AIC(ml)

[Package maxLik version 1.5-2.2 Index]