maxValue {maxLik}	R Documentation

Function value at maximum

Description

Returns the function value at (estimated) maximum.

Usage

maxValue(x, ...)
## S3 method for class 'maxim'
maxValue(x, ...)

Arguments

x
a statistical model, or a result of maximisation, created by maxLik, maxNR or another optimizer.

...
further arguments for other methods

Value

numeric, the value of the objective function at maximum. In general, it is the last calculated value in case the process did not converge.

Author(s)

Ott Toomet

See Also

maxLik, maxNR

Examples

## Estimate the exponential distribution parameter:
t <- rexp(100, 2)
loglik <- function(theta) sum(log(theta) - theta*t)
## Estimate with numeric gradient and numeric Hessian
a <- maxNR(loglik, start=1)
maxValue(a)

[Package maxLik version 1.5-2.2 Index]