summary.maxLik {maxLik}	R Documentation

summary the Maximum-Likelihood estimation

Description

Summary the Maximum-Likelihood estimation including standard errors and t-values.

Usage

## S3 method for class 'maxLik'
summary(object, eigentol=1e-12, ... )
## S3 method for class 'summary.maxLik'
coef(object, ...)

Arguments

object
object of class 'maxLik', or 'summary.maxLik', usually a result from Maximum-Likelihood estimation.

eigentol
The standard errors are only calculated if the ratio of the smallest and largest eigenvalue of the Hessian matrix is less than “eigentol”. Otherwise the Hessian is treated as singular.

...
currently not used.

Value

An object of class 'summary.maxLik' with following components:

type

type of maximization.

iterations

number of iterations.

code

code of success.

message

a short message describing the code.

loglik

the loglik value in the maximum.

estimate

numeric matrix, the first column contains the parameter estimates, the second the standard errors, third t-values and fourth corresponding probabilities.

fixed

logical vector, which parameters are treated as constants.

NActivePar

number of free parameters.

constraints

information about the constrained optimization. Passed directly further from maxim-object. NULL if unconstrained maximization.

Author(s)

Ott Toomet, Arne Henningsen

See Also

maxLik for maximum likelihood estimation, confint for confidence intervals, and tidy and glance for alternative quick summaries of the ML results.

Examples

## ML estimation of exponential distribution:
t <- rexp(100, 2)
loglik <- function(theta) log(theta) - theta*t
gradlik <- function(theta) 1/theta - t
hesslik <- function(theta) -100/theta^2
## Estimate with numeric gradient and hessian
a <- maxLik(loglik, start=1, control=list(printLevel=2))
summary(a)
## Estimate with analytic gradient and hessian
a <- maxLik(loglik, gradlik, hesslik, start=1, control=list(printLevel=2))
summary(a)

[Package maxLik version 1.5-2.2 Index]